cafe/requirements.txt
django~=3.7
django-filter
djangorestframework
